Always include dune and always exclude threads
Any dune project should have a dependency on dune.

Also, dune lists `threads` as an external library you need, even though
in fact it has its own built-in version. So filter that out.

Also, only consider a package as owning an ocamlfind library if it added
the `META` file, not just if it added any file to its directory. Not
sure if this makes any difference, but might be useful.

Fixes #1 and #2.