Use "dune describe" to find internal libraries
Avoids warning about missing internal dependencies.

Also, skip dependencies of vendored projects; assume that the vendored
projects themselves will handle their dependencies correctly.

Note that this assumes we're using vendoring as a way of pinning
unreleased software. It's possible that people might also use vendoring
as a way to include a library privately. In that case, this isn't what
is wanted and we might need to add an option later.