Replace "dune external-lib-deps" command
The new version of dune (from 3.0.0) removed "dune external-lib-deps" and
this is a patch that work with a patch of dune from
"https://github.com/moyodiallo/dune/tree/opam-dune-lint".

There's already a PR(https://github.com/ocaml/dune/pull/6045) in dune
about this patch, but not guarantee to be merged. The command may change also.