When an executable is copied before installed in dune
This case is a hack to make dependency on an executable. It's trying to
copy in another name before to install it. OCaml-CI raised this issue
which can be found here `https://github.com/ocaml/dune/issues/3499`

This commit try solve the case but not perfect because to get the rules,
we are parsing the dune source and there's some limitation on that, like
dune variables.